Extract ReuseObjectPool
